App-specific patching *seems* to be working properly now.
There was an issue earlier where all files were copied to the one target.  It
would probably have done the same thing with multiple targets, but I've not
checked it yet.  Now it copies only the application's files to APPDIR\patches\new,
and patches the application as usual from there.

Cleanup commit to follow.