Okay, now it's really working. ;)
Last commit still had the old ChannelManager registry data hanging around.
That's commented out and will be removed in a separate commit.  I'm
leaving it in for now so that it stands out when it's removed.  (This commit
should have been only for cleaning up old comments.)

Finally, the patch.manifest is not needed and should now be removed.